chore(code-style): Configure ESLint
- Configure ESLint
- Set up pre-commit hooks to flag style violations
- Setup babel transpilation
- Fix npm vulns

[Finishes #164139761]
